Cleaning up the What's Open Swift files. Putting all of the 3 enums in a WhatIsOpen Enum grouping. Moving the 2 extensions to the extensions group. Moving the Mock Data to a file in the UnitTesting group.
